+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- [Library] Improve unit tests for border raw tokens
+- [Library] Improve unit tests for dimensions raw tokens
+- [Library] Improve unit tests for borders raw tokens
 
 ### Fixed